Split GUI sample browser tests
